Make this library compile on non-Linux systems (#6)
* Make this library compile on non-Linux systems

Clarify that we fully support Linux only. On non-Linux systems
this code will not behave as intended.

However, it may still be useful to compile on non-Linux systems
for running smoke testing. (For example, my laptop is a macOS
system and I routinely work on m-lab/ndt-server using it, hence
it's useful for me that this code works on macOS.)

There is no point in sweating to make this code work well for
non-Linux systems, or to make tests pass, because M-Lab is
a Linux based platform. So, this diff is only for convenience.

* Update README.md

* Repair build on Linux systems. Ugh