Uses POD_NAME env instead of os.Hostname()
Currently, the prefix package uses the return value of os.Hostname() as
a prefix for UUIDs. But we have discovered that the value of hostname
can sometimes differ in containers depending on the environment (e.g.,
physical vs virtual). The original intent was that the value of
os.Hostname() in a container would always be the pod name. This commit
leverages a new environment variable inside the uuid initContainer which
will explicitly have the pod's name instead of relying on os.Hostname()
may return.