Adds explict tests for POD_NAME and os.Hostname()
Previously, the unit test was not correctly checking for when the UUID
prefix was being sourced from the env variable POD_NAME, or was falling
back to os.Hostname(). This commit adds a new test to check this, and
updates an existing check to be explicit about where the return value of
generate() is coming from.